chore: set up top level exports for @cdktf/hcl2json
